moved code for sharing items to a helper class
