+++ src/Files.App/Actions/Content/Share/ShareItemAction.cs
@@ -3,16 +3,10 @@ using CommunityToolkit.Mvvm.DependencyInjection;
 using Files.App.Commands;
 using Files.App.Contexts;
 using Files.App.Extensions;
-using Files.App.Filesystem;
-using Files.App.Filesystem.StorageItems;
 using Files.App.Helpers;
-using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Windows.ApplicationModel.DataTransfer;
-using Windows.Foundation;
-using Windows.Storage;
 
 namespace Files.App.Actions
 {
@@ -29,10 +23,7 @@ namespace Files.App.Actions
 		public bool IsExecutable => IsContextPageTypeAdaptedToCommand() &&
 			DataTransferManager.IsSupported() &&
 			context.SelectedItems.Any() &&
-			!context.SelectedItems.Any(i =>
-				i.IsHiddenItem ||
-				(i.IsShortcut && !i.IsLinkItem) ||
-				(i.PrimaryItemAttribute == StorageItemTypes.Folder && !i.IsArchive));
+			context.SelectedItems.All(ShareItemHelpers.IsItemShareable);
 
 		public ShareItemAction()
 		{
@@ -41,72 +32,7 @@ namespace Files.App.Actions
 
 		public Task ExecuteAsync()
 		{
-			var interop = DataTransferManager.As<UWPToWinAppSDKUpgradeHelpers.IDataTransferManagerInterop>();
-			IntPtr result = interop.GetForWindow(App.WindowHandle, UWPToWinAppSDKUpgradeHelpers.InteropHelpers.DataTransferManagerInteropIID);
-
-			var manager = WinRT.MarshalInterface<DataTransferManager>.FromAbi(result);
-			manager.DataRequested += new TypedEventHandler<DataTransferManager, DataRequestedEventArgs>(Manager_DataRequested);
-
-			interop.ShowShareUIForWindow(App.WindowHandle);
-
-			async void Manager_DataRequested(DataTransferManager sender, DataRequestedEventArgs args)
-			{
-				DataRequestDeferral dataRequestDeferral = args.Request.GetDeferral();
-				List<IStorageItem> items = new();
-				DataRequest dataRequest = args.Request;
-
-				foreach (ListedItem item in context.SelectedItems)
-				{
-					if (item is ShortcutItem shItem)
-					{
-						if (shItem.IsLinkItem && !string.IsNullOrEmpty(shItem.TargetPath))
-						{
-							dataRequest.Data.Properties.Title = string.Format("ShareDialogTitle".GetLocalizedResource(), item.Name);
-							dataRequest.Data.Properties.Description = "ShareDialogSingleItemDescription".GetLocalizedResource();
-							dataRequest.Data.SetWebLink(new Uri(shItem.TargetPath));
-							dataRequestDeferral.Complete();
-
-							return;
-						}
-					}
-					else if (item.PrimaryItemAttribute == StorageItemTypes.Folder && !item.IsArchive)
-					{
-						if (await StorageHelpers.ToStorageItem<BaseStorageFolder>(item.ItemPath) is BaseStorageFolder folder)
-							items.Add(folder);
-					}
-					else
-					{
-						if (await StorageHelpers.ToStorageItem<BaseStorageFile>(item.ItemPath) is BaseStorageFile file)
-							items.Add(file);
-					}
-				}
-
-				if (items.Count == 1)
-				{
-					dataRequest.Data.Properties.Title = string.Format("ShareDialogTitle".GetLocalizedResource(), items.First().Name);
-					dataRequest.Data.Properties.Description = "ShareDialogSingleItemDescription".GetLocalizedResource();
-				}
-				else if (items.Count == 0)
-				{
-					dataRequest.FailWithDisplayText("ShareDialogFailMessage".GetLocalizedResource());
-					dataRequestDeferral.Complete();
-
-					return;
-				}
-				else
-				{
-					dataRequest.Data.Properties.Title = string.Format(
-						"ShareDialogTitleMultipleItems".GetLocalizedResource(),
-						items.Count,
-						"ItemsCount.Text".GetLocalizedResource());
-					dataRequest.Data.Properties.Description = "ShareDialogMultipleItemsDescription".GetLocalizedResource();
-				}
-
-				dataRequest.Data.SetStorageItems(items, false);
-				dataRequestDeferral.Complete();
-
-				// TODO: Unhook the event somewhere
-			}
+			ShareItemHelpers.ShareItems(context.SelectedItems.ToArray());
 
 			return Task.CompletedTask;
 		}
